fix(restore): backup intro screens seen
