+++ src/store/slices/settings.ts
@@ -14,6 +14,7 @@ import { EAvailableNetwork } from '../../utils/networks';
 import { EDenomination, EUnit } from '../types/wallet';
 
 export type TSettings = {
+	backupVerified: boolean;
 	enableAutoReadClipboard: boolean;
 	enableSendAmountWarning: boolean;
 	enableSwipeToHideBalance: boolean;
@@ -47,6 +48,10 @@ export type TSettings = {
 	treasureChests: TChest[];
 	orangeTickets: string[];
 	webRelay: string;
+	quickpayIntroSeen: boolean;
+	transferIntroSeen: boolean;
+	spendingIntroSeen: boolean;
+	savingsIntroSeen: boolean;
 };
 
 export const settingsSlice = createSlice({
@@ -83,6 +88,9 @@ export const settingsSlice = createSlice({
 		addOrangeTicket: (state, action: PayloadAction<string>) => {
 			state.orangeTickets.push(action.payload);
 		},
+		verifyBackup: (state) => {
+			state.backupVerified = true;
+		},
 		resetSettingsState: () => initialSettingsState,
 	},
 });
@@ -95,6 +103,7 @@ export const {
 	addTreasureChest,
 	updateTreasureChest,
 	addOrangeTicket,
+	verifyBackup,
 	resetSettingsState,
 } = actions;
