+++ src/store/slices/user.ts
@@ -1,43 +1,35 @@
+// NOTE: 'user' slice is persisted to storage, but not included in remote backup
+
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 
 export const MAX_WARNINGS = 3; // how many times to show high balance warning
 
 export type TUser = {
-	backupVerified: boolean;
 	ignoreAppUpdateTimestamp: number;
 	ignoreBackupTimestamp: number;
 	ignoreHighBalanceCount: number;
 	ignoreHighBalanceTimestamp: number;
 	isGeoBlocked: boolean;
 	lightningSettingUpStep: number;
-	quickpayIntroSeen: boolean;
 	requiresRemoteRestore: boolean;
 	startCoopCloseTimestamp: number;
 	ignoresHideBalanceToast: boolean;
 	ignoresSwitchUnitToast: boolean;
 	scanAllAddressesTimestamp: number;
-	transferIntroSeen: boolean;
-	spendingIntroSeen: boolean;
-	savingsIntroSeen: boolean;
 };
 
 export const initialUserState: TUser = {
-	backupVerified: false,
 	ignoreAppUpdateTimestamp: 0,
 	ignoreBackupTimestamp: 0,
 	ignoreHighBalanceCount: 0,
 	ignoreHighBalanceTimestamp: 0,
 	isGeoBlocked: false,
 	lightningSettingUpStep: 0,
-	quickpayIntroSeen: false,
 	requiresRemoteRestore: false,
 	startCoopCloseTimestamp: 0,
 	ignoresHideBalanceToast: false,
 	ignoresSwitchUnitToast: false,
 	scanAllAddressesTimestamp: 0,
-	transferIntroSeen: false,
-	spendingIntroSeen: false,
-	savingsIntroSeen: false,
 };
 
 export const userSlice = createSlice({
@@ -67,9 +59,6 @@ export const userSlice = createSlice({
 		clearCoopCloseTimer: (state) => {
 			state.startCoopCloseTimestamp = 0;
 		},
-		verifyBackup: (state) => {
-			state.backupVerified = true;
-		},
 		ignoreHideBalanceToast: (state) => {
 			state.ignoresHideBalanceToast = true;
 		},
@@ -90,7 +79,6 @@ export const {
 	setLightningSetupStep,
 	startCoopCloseTimer,
 	clearCoopCloseTimer,
-	verifyBackup,
 	ignoreHideBalanceToast,
 	ignoreSwitchUnitToast,
 	resetUserState,
